// LibrarySystem_FrontV1.1/src/app/models/personinfomodel.ts
    export class PersonInfoModel {
        id: string;
        isDelete: number;
        createTime: number;
        operator: string;
        updateTime: number;
        accountAddress: string;
        code: string;
        remark: string;
        entranceDate: number;
        graduateSchool: string;
        identityCardId: string;
        nativePlacenativePlace: string;
        pictrue: string;
        politicalStatus: string;
        sex: string;
        status: string;
        telNum: string;
        isGraduate?: any;
        name: string;
        nation: string;
        classesName: string;
        professionName: string;
        email: string;
        noGraduateReason: string;
        originalPlace: string;
        usedName: string;
        isRegistry?: any;
        payNumber: string;
        roomId: string;
        alluserid:string;
    }